AI/requirements.txt
flask
flask-cors
sentence-transformers
chromadb
pypdf
langchain
langchain-community
langchain-text-splitters
python-dotenv
tqdm
google-generativeai
